Implement central place for models
