+++ convex/maps.ts
@@ -1,5 +1,12 @@
+import OpenAI from "openai";
+import { internalAction, internalMutation, query } from "./_generated/server";
 import { v } from "convex/values";
-import { internalMutation, query } from "./_generated/server";
+import { z } from "zod";
+import { zodResponseFormat } from "openai/helpers/zod";
+import { Doc } from "./_generated/dataModel";
+import { ZombieSurvival } from "../simulators/zombie-survival";
+import { api, internal } from "./_generated/api";
+import { runModel } from "../models";
 
 const MAPS = [
   {
@@ -96,5 +103,79 @@ export const getMapByLevel = query({
       .query("maps")
       .withIndex("by_level", (q) => q.eq("level", args.level))
       .first();
+  },
+});
+
+export const playMapAction = internalAction({
+  args: {
+    level: v.number(),
+    gameId: v.id("games"),
+    modelId: v.string(),
+  },
+  handler: async (ctx, args) => {
+    const resultId = await ctx.runMutation(
+      internal.results.createInitialResult,
+      {
+        gameId: args.gameId,
+        level: args.level,
+      },
+    );
+
+    const map: Doc<"maps"> | null = (await ctx.runQuery(
+      api.maps.getMapByLevel,
+      {
+        level: args.level,
+      },
+    )) as any;
+
+    if (!map) {
+      throw new Error("Map not found");
+    }
+
+    if (process.env.MOCK_MODELS === "true") {
+      const existingMap = [...map.grid.map((row) => [...row])];
+
+      existingMap[0][0] = "P";
+      existingMap[0][1] = "B";
+      existingMap[0][2] = "B";
+
+      const game = new ZombieSurvival(existingMap);
+
+      while (!game.finished()) {
+        game.step();
+      }
+
+      const isWin = !game.getPlayer().dead();
+
+      await ctx.runMutation(internal.results.updateResult, {
+        resultId,
+        isWin,
+        reasoning: "This is a mock response",
+        map: existingMap,
+      });
+
+      return;
+    }
+
+    try {
+      const { solution, reasoning } = await runModel(args.modelId, map.grid);
+      const game = new ZombieSurvival(solution);
+
+      while (!game.finished()) {
+        game.step();
+      }
+
+      const isWin = !game.getPlayer().dead();
+
+      await ctx.runMutation(internal.results.updateResult, {
+        resultId,
+        isWin,
+        reasoning,
+        map: solution,
+      });
+    } catch (error) {
+      // todo: handle error
+      console.log(error);
+    }
   },
 });
